backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pandas>=2.2.0
numpy==1.26.2
scikit-learn==1.3.2
shap==0.43.0
requests==2.31.0
python-dotenv==1.0.0
joblib==1.3.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
sentence-transformers>=2.2.0
beautifulsoup4>=4.12.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
openai>=1.0.0
anthropic>=0.39.0
reportlab>=4.0.0
